Move issue capping to validate-issue-logging
